Adds support for linking Foes to Battles
Through Denormalization, Foes can be linked to Battles which will ultimately affect the Battle's stats.

This is done through a new helper called Pathfinder which contains a small set of the rules for determining CR, XP, and Treasure.

Because "Linking" is a bit more involved, it's on a Model by Model basis for now.  It's, essentially, a denormalized one-to-many relationship.