Add Encounters to Bounty, Task, Area, and Shop
Last major piece of the Manage component of Dungeon Keeper (yay!)

This adds Encounters.  Encounters are essentially the bread-and-butter of a campaign.  They can range from Traps to Combat to Social interactions.
Everything else to this point has been created purely for organization, but Encounters are 100% the "meat" of a Campaign.

At this point, I can fully create a Pathfinder Campaign using the interface.

Next step:  The much simpler Adventure Mode.