AttrBlock is Awesome!
AttrBlock always had so much potential, but it was squandered basically serving as a display component for single model values.

But no more.

AttrBlock is now a full-featured display component - offering pre-designed templates, class orders, styles, and variables that give the entire UI a consistency not achieved before.

With the ability to add even more templates as time goes on, AttrBlock is now one of the most used and important components I have.